(unifying) move Ethereum types into one file
