+++ src/js/types/ethereum.js
@@ -2,6 +2,14 @@
 // Ethereum types
 // https://github.com/ethereumjs/ethereumjs-tx
 
+import type { $Path, $Common } from './params';
+import type { Unsuccessful$ } from './response';
+import type {
+    Success,
+    MessageSignature,
+    EthereumSignedTx,
+} from './trezor';
+
 export type Transaction = {
     to: string,
     value: string,
@@ -15,3 +23,64 @@ export type Transaction = {
     r: string,
     s: string,
 }
+
+// get address
+
+export type EthereumAddress = {
+    address: string,
+    path: Array<number>,
+    serializedPath: string,
+}
+
+export type $EthereumGetAddress = {
+    path: $Path,
+    showOnTrezor?: boolean,
+}
+
+export type EthereumGetAddress$ = {
+    success: true,
+    payload: EthereumAddress,
+} | Unsuccessful$;
+
+export type EthereumGetAddress$$ = {
+    success: true,
+    payload: Array<EthereumAddress>,
+} | Unsuccessful$;
+
+// sign transaction
+
+export type $EthereumSignTransaction = $Common & {
+    path: $Path,
+    transaction: Transaction,
+}
+
+export type EthereumSignTransaction$ = {
+    success: true,
+    payload: EthereumSignedTx,
+} | Unsuccessful$;
+
+// sign message
+
+export type $EthereumSignMessage = $Common & {
+    path: $Path,
+    message: string,
+}
+
+export type EthereumSignMessage$ = {
+    success: true,
+    payload: MessageSignature,
+} | Unsuccessful$;
+
+// verify message
+
+export type $EthereumVerifyMessage = $Common & {
+    address: string,
+    message: string,
+    signature: string,
+}
+
+export type EthereumVerifyMessage$ = {
+    success: true,
+    payload: Success,
+} | Unsuccessful$;
+
